Spring security using JWT

+++ src/main/java/com/springboot/projects/airBnbApp/service/RoomServiceImpl.java
@@ -3,13 +3,16 @@ package com.springboot.projects.airBnbApp.service;
 import com.springboot.projects.airBnbApp.dto.RoomDto;
 import com.springboot.projects.airBnbApp.entity.Hotel;
 import com.springboot.projects.airBnbApp.entity.Room;
+import com.springboot.projects.airBnbApp.entity.User;
 import com.springboot.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.springboot.projects.airBnbApp.exception.UnauthorizedException;
 import com.springboot.projects.airBnbApp.repository.HotelRepository;
 import com.springboot.projects.airBnbApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,6 +33,10 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID : "+hotelId));
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel hence cant modify rooms: "+user.getId());
+        }
         room = roomRepository.save(room);
         // TO DO : To add to inventory as soon as DB updates with new Room
         inventoryService.addInventory(room);
@@ -61,6 +68,13 @@ public class RoomServiceImpl implements RoomService{
         if(!exists){
             throw new ResourceNotFoundException("Room not found with Id : "+ roomId);
         }
+
+        Room room = roomRepository.findById(roomId).orElseThrow(()-> new ResourceNotFoundException("Room not found : "+roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel : "+user.getId());
+        }
+
         inventoryService.deleteRoomsByRoomId(roomId);
         roomRepository.deleteById(roomId);
     }
